rename getXvsVaultPoolLength to getXvsVaultPoolsCount
