+++ src/clients/api/__mocks__/index.ts
@@ -79,9 +79,9 @@ export const getTransactions = jest.fn();
 export const useGetTransactions = () =>
   useQuery([FunctionKey.GET_TRANSACTIONS, {}], getTransactions);
 
-export const getXvsVaultPoolLength = jest.fn();
-export const useGetXvsVaultPoolLength = () =>
-  useQuery(FunctionKey.GET_XVS_VAULT_POOL_LENGTH, getXvsVaultPoolLength);
+export const getXvsVaultPoolsCount = jest.fn();
+export const useGetXvsVaultPoolsCount = () =>
+  useQuery(FunctionKey.GET_XVS_VAULT_POOL_LENGTH, getXvsVaultPoolsCount);
 
 export const getXvsVaultPoolInfos = jest.fn();
