+++ src/clients/api/queries/getXvsVaultPoolsCount.spec.ts
@@ -1,10 +1,10 @@
 import { XvsVault } from 'types/contracts';
 import { TOKENS } from 'constants/tokens';
-import getXvsVaultPoolLength, { GetXvsVaultPoolLengthOutput } from './getXvsVaultPoolLength';
+import getXvsVaultPoolsCount, { GetXvsVaultPoolsCountOutput } from './getXvsVaultPoolsCount';
 
 const xvsTokenAddress = TOKENS.xvs.address;
 
-describe('api/queries/getXvsVaultPoolLength', () => {
+describe('api/queries/getXvsVaultPoolsCount', () => {
   test('throws an error when request fails', async () => {
     const fakeContract = {
       methods: {
@@ -17,18 +17,18 @@ describe('api/queries/getXvsVaultPoolLength', () => {
     } as unknown as XvsVault;
 
     try {
-      await getXvsVaultPoolLength({
+      await getXvsVaultPoolsCount({
         xvsVaultContract: fakeContract,
       });
 
-      throw new Error('getXvsVaultPoolLength should have thrown an error but did not');
+      throw new Error('getXvsVaultPoolsCount should have thrown an error but did not');
     } catch (error) {
       expect(error).toMatchInlineSnapshot('[Error: Fake error message]');
     }
   });
 
   test('returns the XVS vault pool length on success', async () => {
-    const fakeOutput: GetXvsVaultPoolLengthOutput = 10;
+    const fakeOutput: GetXvsVaultPoolsCountOutput = 10;
 
     const callMock = jest.fn(async () => fakeOutput);
     const poolLengthMock = jest.fn(() => ({
@@ -41,7 +41,7 @@ describe('api/queries/getXvsVaultPoolLength', () => {
       },
     } as unknown as XvsVault;
 
-    const response = await getXvsVaultPoolLength({
+    const response = await getXvsVaultPoolsCount({
       xvsVaultContract: fakeContract,
     });
